Make MutableCSSPropertyValueSet into an enum.
Currently, this is a struct with two flags (did_parse, did_change),
but only three of the combinations are actually possible (a value that
didn't parse, but changed something, is meaningless). As the next patch
will introduce a fourth state, change it into an enum instead of
introducing a third non-orthogonal flag.

Change-Id: I93db46c8530dc9957938a6c6ea077ec0c7187c13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3449268
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#976614}
NOKEYCHECK=True
GitOrigin-RevId: 5636e209a7a8a839555b33b8dfbd1066269e152d