Remove kinuko@chromium.org from third_party/blink/public/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ib9d5b4fd9c578768dab996eb026f08ede9806b41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3499277
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977172}
NOKEYCHECK=True
GitOrigin-RevId: 7f162507bde89ec2b6b0c5ae07e35a71bc361084