Remove kinuko@chromium.org from third_party/blink/renderer/platform/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I566872a8c10566bc5063e403c4eb8fa676ebd3da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501958
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977176}
NOKEYCHECK=True
GitOrigin-RevId: fe8875a432623828f06c2fe2dbf999e759cea3ea