Remove kinuko@chromium.org from third_party/blink/renderer/core/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I1a12e53a0857812049cd961029bb3ac3920211e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501505
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977177}
NOKEYCHECK=True
GitOrigin-RevId: 03c7eb742a5beca51121b86896a4289a7692b220