Remove rsleevi@chromium.org from third_party/google_trust_services/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I9a5c69a3e85862951f884c832a86a6a79cb562ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3502090
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977214}
NOKEYCHECK=True
GitOrigin-RevId: 66b41cf85abaaf071bebfdc03c36b5371489c1f6