Remove aboxhall@chromium.org from third_party/blink/renderer/modules/accessibility/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ia00d3735724b5d4e3cd0d95d97c3b0fe36d7d68e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501547
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977237}
NOKEYCHECK=True
GitOrigin-RevId: 1e9f2ac67ed53d94b85c4e02958b6d3152f1b858