Remove kinuko@chromium.org from third_party/blink/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ibe26bcf4158692292fe4c16246c9ad68b15abd26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501988
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Rick Byers <rbyers@chromium.org>
Commit-Queue: Rick Byers <rbyers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977284}
NOKEYCHECK=True
GitOrigin-RevId: db7818b8da8d40807fb4ca407f3a91d5a1bd4ae4