[FedCM] Make privacy policy url required
Currently we allow no privacy policy url to be specified in the client
metadata. This CL changes that by surfacing an error if the JSON parser
returns an empty value for |privacy_policy_url|. We also sanity check
the privacy policy, and if it is an invalid URL then reject. In order to
properly surface the problem about missing privacy policy, we add a
specific console error message. We also add a new value to the
FederatedAuthRequestIssueReason so that a similar message can be
displayed in the Issues panel.

Bug: 1293913
Change-Id: I5dbb6e6a1005db33dccafb7977fb7cced8f871d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3499232
Reviewed-by: Yi Gu <yigu@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Nicolás Peña <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977304}
NOKEYCHECK=True
GitOrigin-RevId: 09408255f5fc40e54ce586e5f3e4048dd87a6f58