Remove aboxhall@chromium.org from third_party/iaccessible2/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ib9ea140cc5d4ff54ab241c567526f3c95762ef90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501962
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977317}
NOKEYCHECK=True
GitOrigin-RevId: 67e7d784f76a29545c3d79c7f21ecac9f3da3382