[css-flex] Send layout data to devtools
In this attempt, we mark the flex layout object as needing layout for
devtools and re-run the rendering pipeline. When needing layout for
devtools, the flex algorithm leaves the devtools info on the layout
result, which devtools then reads.

I'm leaving the existing devtools code for recreating this info from the
fragment because I've had to revert this functionality so many times, I
want to leave that as backup.

Change-Id: Ib988107a4c433a7ad95c0af888689ffef7f8d425
Bug: 1186072
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3496750
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977503}
NOKEYCHECK=True
GitOrigin-RevId: 2a2587b25174a4c500614e7fec6663d057cd5c73