[TestExpectations] Disable layout tests affected by DevTools change
This CL disables a range of CLs that are affected by a DevTools
change: crrev.com/c/3470237.

Some of these tests are unrelated to the exact change, but
change the timing of things that happen and require a change
of the test suite.

Others are related to this change, and will be addressed.

Bug: chromium:1300509, chromium:1229541, chromium:1133307
Change-Id: I17995bd7b56bae0140e58ab25a48d8fcab79f2a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3487867
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977632}
NOKEYCHECK=True
GitOrigin-RevId: 4cabfc4491638b17865aa9884001d0912d3d23b2