Remove majidvp@chromium.org from third_party/blink/renderer/modules/animationworklet/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I750695af3878220bb58e350966751cf851657d8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501954
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977648}
NOKEYCHECK=True
GitOrigin-RevId: 621f59ac24911ae364f1270f1cc97e5b6624caf8