Use attributionsrc infra in registerAttributionSource JS method
We replace the AttributionSourceParams struct with a string parameter
representing an attributionsrc URL to be queried the same way as the
corresponding attributionsrc HTML attribute. We reuse the
AttributionSrcLoader::Register method for this.

https://github.com/WICG/conversion-measurement-api/blob/main/EVENT.md#registering-attribution-sources

Note that this API was and is only exposed behind the test-only
ConversionMeasurementEventSources runtime enabled feature. The OT for
the main ConversionMeasurement feature ended in M98.

This change also allows us to remove the RegisterImpression Mojo method
from the ConversionHost interface and implementation, along with
associated infrastructure and tests.

Bug: 1302622, 1302680, 1275201, 1301960
Change-Id: If39b412facf5887175c9570255332d88c8db1b80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3499035
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977983}
NOKEYCHECK=True
GitOrigin-RevId: d1fc30856711727847967cab405c3785b839f091