Fix gesture related Fenced Frame WPT flakiness
The gesture API of WPT Test Driver doesn't work well in MPArch Fenced
Frame on Mac. The failure happens when the window size is not correctly
initialized. So this CL makes simulateGesture() in utils.js and the code
in hid.https.html wait until window.innerWidth is set.

And also this CL makes simulateGesture() to an async function for code
readability.

Bug: 1275997
Change-Id: Ic52d602f35a8fe328622b4daf27268dc31f0c3a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3494985
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@google.com>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978188}
NOKEYCHECK=True
GitOrigin-RevId: b9f354d2120b6cc7b5a5cea2acd639c59310439d