Forms: Change function as data member of CalendarNavigationButton
Currently we pass |onMouseDown| function to add/removeEventListener
using bind(this).
But that doesn't work as expected because bind(this) creates new object
every time.
This CL adds new bound function as a data member so that it can be used.

Bug: 981231
Change-Id: I1b8058f9daff4d24b137787553bdf2231d089304
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3505988
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Jihwan Kim <bluewhale.marc@gmail.com>
Cr-Commit-Position: refs/heads/main@{#978201}
NOKEYCHECK=True
GitOrigin-RevId: c451679da96a80743abbdd68085f1539989ee3a7