[locals] Build all locale resources
This CL replaces |locales_with_pseudolocales| with |all_chrome_locales|
in most grit targets (remoting and fuchsia targets are ignored for now).

Doing this means that outputs for all pak file outputs need to be added
to the header of all affected grd files. These resources are built but
only those in |platform_pak_locales| are actually packed into each
platforms binary.

The following GRD files are exempt from this CL and will still require
updating if the locales for the platform change:
 - chrome/app/resources/locale_settings_chromiumos.grd
 - chrome/app/resources/locale_settings_google_chromeos.grd
 - chrome/app/resources/locale_settings_linux.grd
 - chrome/app/resources/locale_settings_mac.grd
 - chrome/app/resources/locale_settings_win.grd
 - ios/chrome/app/strings/ios_strings.grd

This is done to make sure that |platform_pak_locales| matches what is
expected for each platform. This still greatly simplifies the grd build
and makes it easier to add more locales to individual platforms.

This CL makes no changes to the locales that are actually shipped on
each platform. Some extra work is added to the build for the unshipped
locales, but looking at bot build times this is not noticeably
significant.

The only string changes done in this CL are removing whitespace at
the end of messages.  EOL whitespace in the middle of messages is not
removed so that new message ids are not created.

Bug: 1293908
Change-Id: I252e0ab410676b15ebb9849ae8c770dcb050abf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3498882
Reviewed-by: Michael Cui <mlcui@google.com>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Owners-Override: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Trevor Perrier <perrier@chromium.org>
Reviewed-by: Claudio M <claudiomagni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978496}
NOKEYCHECK=True
GitOrigin-RevId: 0ab73d8ef011bd36076a1b07698e3cb63bacc52d