Fenced Frames: Add sandboxed frame adoption test
This CL introduces a test wherein we create a fenced frame in a document
not suitable for a fenced frame, and then adopt the frame into a
document that is suitable for a fenced frame. The test then asserts that
the element is fully functional and navigable after the adoption.

This CL also adds TestExpectations to cater to the fact that our
implementation does not yet leave the fenced frame in a functional state
in this circumstance. This will be fixed in crrev.com/c/3501941 as a
part of
https://docs.google.com/document/d/1rbevNYwcX4jsr3a3L-1W96ffWNpGHSCJ0De9KzUQ4Eg/edit.

R=linnan@chromium.org

Bug: 1123606
Change-Id: I66dd0f84e7e7bcc3f46d33fb773ef2c4a1696ecd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3507474
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978576}
NOKEYCHECK=True
GitOrigin-RevId: 7e3ad3bb0605bc2fa505c2fa667fbad1ced65d11