Correct process and thread priorities for lacros
This change will reset the lacros browser main thread priority to normal
before more processes are launched.

It also enables features for lacros that force display priority for
certain threads. Those features are already on for ash, this turns them
on for lacros also.

The impact of this is as follows:
1. It forces the use of the correct cpuset cgroups (chrome,
   chrome/urgent, chrome/non-urgent).

2. It correctly sets the latency_sensitive settings for uclamp.

3. It correctly sets the schedtune cgroups.

4. It correctly sets the thread nice values (priority).

Bug: 217185330
Change-Id: I188228c5ffb158abce12d106219d0c16d381a247
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3498696
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Youssef Esmat <youssefesmat@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978581}
NOKEYCHECK=True
GitOrigin-RevId: fd93a6750f22ca043c7dda62c3750dc3d9941ae4