[@container] Implicit container selection, remove name()/type()
There was a resolution recently that (nameless) container queries
should automatically evaluate against the nearest container that
supports all the features used in the query. For example, given a
query ((width:1px) and (height:1px)), and supposing that the nearest
container has container-type:inline-size, the query should not fail,
but instead keep searching the ancestor chain for a container with
container-type:size.

This CL implements that behavior, and removes the name() and type()
functions from the @container prelude.

Inspector: Changed FindContainerForElement to accept a
ContainerSelector instead of a name. This is because a real
ContainerSelector is required to get accurate container selection
given the new behavior, and we had a test affected by this.
(The element with class "query-unnamed" will now through implicit
container selection reach beyond the immediate named container, and
query the outer unnamed container, hence that element is marked with
"desc-unnamed").

Fixed: 1300545, 1300547
Change-Id: Ic97ff34ecc1f49caa33940ca1451af21ac94d20b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3497355
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978735}
NOKEYCHECK=True
GitOrigin-RevId: b4c16f9f9050c6848f6ea74e74280ff44baf4bdc