CodeHealth: Use Value::Dict in InspectorFrontendClient::SendMessageToEmbedder
The message sent to the embedder is expected to be a dictionary value,
so this converts SendMessageToEmbedder and related methods to use the
v2 base::Value::Dict APIs. The message is ultimately sent using mojo and
a mojo_base.mojom.DictionaryValue, and we convert to a base::Value at
that boundary (which is covered by crbug.com/1301890).

Bug: 1303949

Change-Id: I1a5999326ac1ffc06bcfbe6c2216533a82ea7c9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3504099
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978774}
NOKEYCHECK=True
GitOrigin-RevId: 7f6ec029290879227af3e3dcb85ca63428cf1224