Allow failed GMB readback to fallthrough to normal readback path.
The GMB readback path can fail for a number of reasons, so we
should allow fallthrough and disable the GMB path on error. We
already do this for the WebCodecs path, so we should for RTC too.

Bug: 1302999
Change-Id: Iedefd17fb76b5ebdf7ce7b08d8bff314a370a619
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3508495
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978802}
NOKEYCHECK=True
GitOrigin-RevId: 8c63ac85cf01031b147cb723c43ba49bca70c5c4