[GridFragmentation] Prevent infinite expansion.
We allow grid-items to expand their grid-rows under some common
conditions. This is good as it allows text to not overlap, etc.

However when combined with a %-block-size descendant, we might end up
in a situation where we'll loop indefinitely trying to constantly
grow the row to the item. E.g.

  <div style="display: grid;">
    <div style="min-height: 100px;">
      <div style="height: 200%; background: green;"></div>
    </div>

This patch disables this behaviour by not setting the
MinBlockSizeShouldEncompassIntrinsicSize flag for any grid item with
a %-height-descendant.

Bug: 614667
Change-Id: Ie52b5fb2903259ed92ce6e5cbc5b200a94f088dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3508504
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978881}
NOKEYCHECK=True
GitOrigin-RevId: 7fce7f61957b31667bd18360caf84ca4425cfd7c