[tracing] Add tracing to understand the source of new frames
Currently the source of a new frame requested by ScheduleAnimation call
(both the document context and the source in code) is hard to see from
a trace. Add an explicit event capturing the necessary metadata to make
it easier to investigate and understand constantly-animating pages.

Bug: b/211052694
Change-Id: I0c2502a0987b39307afb35f4f1f403481cd96c02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3509828
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978885}
NOKEYCHECK=True
GitOrigin-RevId: 8a1e6b69936869ba8b782cf15993ac7720a2448b