CodeHealth: Use v2 Dict and List APIs in PeerConnectionTracker
This changes usage of base::Value with type DICTIONARY to
base::Value::Dict, as well as converts related methods and data to use
the v2 APIs (Dict and List) for better type safety. The stat reports,
which this CL converts to a Value::List, are ultimately plumbed through
mojo using mojo_base.mojom.ListValue. We convert to a base::Value at
that boundary, which will be converted as part of crbug.com/1301890.

Change-Id: I382f1bb146d043677729649b4e54cec11056d87b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3504326
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978891}
NOKEYCHECK=True
GitOrigin-RevId: daf3e4022b1e05d4e8e9639ff69b7959f80b0a02