Avoid Autoplay handling when an AudioContext is already closed
The Autoplay handling routine in AudioContext assumes the associated
Document and ExecutionContext are always valid. This patch adds a
check in the user gesture trigger if the context is already closed
(uninitialized or destroyed) it simply avoids handling of the
Autoplay routine.

Bug: 1291101
Test: Locally confirmed it does not crash ASAN anymore.
Change-Id: I0625ea787983d772601ff5fd6bcb066029238355
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3511827
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#979054}
NOKEYCHECK=True
GitOrigin-RevId: 9110ca64e1f787de261633c0644e029d353f1125