MPArch Fenced Frame: Measure navigation start time in renderer process
Currently, the navigation start timings of Shadow DOM fenced frame and
MPArch fenced frame are differently measured.

When a new fenced frame is inserted to a page, the navigation start time
of Shadow DOM fenced frame is measured when the fenced frame is inserted
to the document in the renderer process. This is the same as a normal
iframe.

But the navigation start time of MPArch fenced frame is not measured in
the renderer process. The navigation start is measured in the browser
process when it receives the FencedFrameOwnerHost::Navigate IPC message.
  FencedFrame::Navigate()
    => Navigator::NavigateFromFrameProxy()
      => NavigationControllerImpl::NavigateFromFrameProxy()
        => CreateNavigationRequestFromLoadParams()
          => navigation_start = base::TimeTicks::Now();
To fix this issue, this CL introduces navigation_start_time argument in
FencedFrameOwnerHost::Navigate IPC. And use it in CommonNavigationParam.

And also when AvoidUnnecessaryBeforeUnloadCheck feature is not enabled,
the start time is overridden by the time of before_unload_end_time
which is measured in blink::LocalFrameMojoHandler::BeforeUnload().
This is because the browser process unnecessarily sends BeforeUnload
IPC message to the renderer process for legacy reasons
(crbug.com/365039).
To fix this issue, this CL adds "!IsFencedFrameRoot()" check for the
run_beforeunload_for_legacy_frame flag in
RenderFrameHostImpl::CheckOrDispatchBeforeUnloadForSubtree().

Bug: 1263341

Change-Id: Ie2382ec5b030f8e9b4cc76003b6ad8c6305ae11f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3244673
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#979090}
NOKEYCHECK=True
GitOrigin-RevId: 167ca64398568235a79f82bfb177f2a3694f2495