[Fixit] Fix secure context bit for shared workers.
This change is gated by a runtime-enabled Blink feature flag, which is
disabled by default and marked as experimental. When the feature is
disabled, a UseCounter is recorded each time a new shared worker is
instantiated and wrongly considered a secure context. This will be used
to determine the compatibility impact of enabling the feature later on.

The previous code tried to determine whether the creator context was
secure by checking the properties of its origin. That approach fails to
account that contexts with secure origins embedded in non-secure
contexts are not secure.

This change pipes the creator context's secure context bit through Mojo
from the browser process to the renderer, then through various layers
on the renderer side into SharedWorkerGlobalScope via
GlobalScopeCreationParams.starter_secure_context.

In order to provide backwards compatibility when the feature is
disabled, the previously-computed bit (whether the creator context has
a secure origin) which was set in starter_secure_context is now
explicitly piped from WebSharedWorkerImpl to SharedWorkerGlobalScope
outside of GlobalScopeCreationParams. This requires the use of a few
temporary code constructs that should be removed once the feature is
enabled by default.

Finally, the analogous code for dedicated workers is shifted from
WorkerOrWorkletGlobalScope to DedicatedWorkerGlobalScope in what should
be a no-op change to work around C++ inheritance subtleties. The move
preserves behavior for dedicated workers, and switches the behavior of
service workers and worklets to always correctly consider whether the
creator context is secure or not. This should not have an impact on
service workers nor worklets, since they can only be instantiated by
secure contexts.

The aforementioned C++ subtleties are that WorkerOrWorkletGlobalScope's
constructor calls SecurityContext::SetSecurityOrigin(), which calls
back into WorkerOrWorkletGlobalScope::HasInsecureContextInAncestors()
to determine whether the context may be secure or not. Overriding
HasInsecureContextInAncestors() in SharedWorkerGlobalScope has no effect
here, since virtual function calls are disabled during construction [1].

Instead, the logic to use the legacy or fixed behavior for dedicated
workers is moved into DedicatedWorkerGlobalScope's constructor, which
captures a copy of GlobalScopeCreationParams.starter_secure_context
in ParsedCreationParams before modifying it and moving the entire
GlobalScopeCreationParams struct into WorkerGlobalScope. Capturing a
copy of the correct bit allows incrementing a UseCounter later.

The exact same construct is used for SharedWorkerGlobalScope.

[1] https://stackoverflow.com/questions/962132/calling-virtual-functions-inside-constructors

Fixed: chromium:780031
Change-Id: I8310b9f783bbde9b294aea0daf117f180887730b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3289256
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Titouan Rigoudy <titouan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#979137}
NOKEYCHECK=True
GitOrigin-RevId: bafedf217b2d832972c1446261963b5f4995539f