[SH] Build prefix, suffix and range by processing available text as needed.
Previously we get all available text for prefix and suffix and then
consume it word by word as needed. However in cases when available
prefix or suffix is too big (such as in the linked bug) we iterate
over a long list of nodes checking each of them to be valid. This
results in long loop which pauses the renderer and delays the context
menu.

In this CL, adding word iterators that will consider next node only
when a new word is needed. This way we will not uselessly iterate
over long list of nodes when we need only 10 words.

Bug: 1260839
Change-Id: I660e5771359e8bf19e759f7db459eccbc7fc6ca0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3481378
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Gayane Petrosyan <gayane@google.com>
Cr-Commit-Position: refs/heads/main@{#979257}
NOKEYCHECK=True
GitOrigin-RevId: 3a22e3a141cf5b72aa16e483b8b0ff3b19403c09