Clear previously propagated float-in-inline break tokens.
The propagated break tokens from floats inside inlines are picked up and
added to the builder at the same time as adding the fragment. However,
if we before doing that decide to discard the fragment and relayout the
node because the BFC offset got resolved, we'd add the break token for
the same float twice afterwards. Find a better place to clear the break
tokens.

Bug: 1303201
Change-Id: Iad695cdb5dc61228d4bdef868f09b3426230ff65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3512652
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#979264}
NOKEYCHECK=True
GitOrigin-RevId: 23c071b95e90a0d283cfce717e43eb1a5a6fb039