[renderblocking] Unify render-blocking stylesheet and font management
This patch unifies the management of render-blocking stylesheets and
other resources (currently fonts only) with the same class,
RenderBlockingResourceManager.

Note that this patch is not a strict refactoring, though it tries to
replicate the old behavior of how StyleEngineContext controls render-
blocking as much as possible -- a stylesheet is render-blocking if it's
inserted before body.

There are some cases where the behaviors differ, most notably in
non-html documents, where the old behavior still makes stylesheets
render-blocking, but as per spec they should just be script-blocking.
The patch follows the spec-compliant behavior.

This makes it easier for future patches to add more types of render-
blocking resources into RenderBlockingResourceManager.

Bug: 1271296
Change-Id: Ia14679962ab990310fd78630ee98d71183983119
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3503857
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#979505}
NOKEYCHECK=True
GitOrigin-RevId: 64b71f7ab2b5b8038d740c5898572b16da3036b4