Rename StartLoadingDynamicSheet to improve clarity
A previous code review pointed out the function name is unclear:

https://chromium-review.googlesource.com/c/chromium/src/+/3503857/3/third_party/blink/renderer/core/svg/svg_style_element.h#71

The function is called only when an @import rule is dynamically added
and starts loading after its owner sheets has already finished loading,
in which case we need to set the state of the owner sheet to pending.
Hence this patch renames the function for better clarity.

Change-Id: I75dcae9810881f0750463448bb2da0c5d762a6f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3510308
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#979509}
NOKEYCHECK=True
GitOrigin-RevId: d07dc359c960a3f27fdd2000d8afecda9e4c2179