Help TextAutosizer get the correct column inline-size.
Previously, if TextAutosizer::DeepestBlockContainingAllText() returned
the LayoutMultiColumnFlowThread, we'd fail the DCHECK in
MultiplierFromBlock(), because it was marked for layout. Make sure that
it's not marked for layout, and also store the column inline-size there,
just like we do in the legacy engine.

Bug: 829028
Change-Id: I329cfc304d8d1cef42ac1d7063ebd6a9300623cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3513356
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#979670}
NOKEYCHECK=True
GitOrigin-RevId: 2cbce823762cc140c3ad4426f4d1e3ad2f37b487