Rebaseline a few Protocol Handler manual tests
The change landed in r887616 we added additional capabilities to the
url module to have a more consistent scheme fragment encoding, adding
the 'fragment percent-encode set' [1]

This has caused that some URLs that have special characters like '<'
or '"' were encoded using '%". This extra "%" will be encoded using
"%25", hence a scheme fragment with '<' will become '%3C" which should
be '%253C' as result of the ProtocolHandler::TranslateUrl function.

The mentioned change performed a rebaseline of several web and unit
tests. However, there are a few manual tests that need to update the
expected results as well.

[1]: https://url.spec.whatwg.org/#fragment-percent-encode-set

Bug: 1304190
Change-Id: I1389b6df1c7bc530ee45e27a12828520c53bbbbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3512616
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Cr-Commit-Position: refs/heads/main@{#979685}
NOKEYCHECK=True
GitOrigin-RevId: 80f434fb80ef5fc6fe866bd79e5ece721a25571c