Modernize resize-corner-tracking-touch test
This patch addresses a few issues with the test. The terminology was a
bit confusing in the test since calling a method named "resize" even
when not performing a resize, and referring to the operations as touch
scroll even when not scrolling.  We now use touchDrag when referring to
a drag operation that could trigger a resize or a scroll.

The failure with percentage based scrolling was due to how
smoothScrollXY was being called. Touchpad and mousewheel scrolling use
the same gesture type. The gestures are differentiated via the
precise_scroll_delta parameters, which was not specified and defaults
to false (mousewheel scroll). This resulted in different scroll amounts
depending on whether percentage based scrolling was being used.

The state of the document was not reset between tests, which means a
failure in one test can lead to inconsistencies in the remaining tests.

Flakes were fairly easy to reproduce prior to the patch. The test is
still failing with scroll-unification enabled, but is otherwise working
well.

Bug: 1095540
Change-Id: Ie4dc71cca23b6b5fd6de598493975ce63b319a70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3514714
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#979749}
NOKEYCHECK=True
GitOrigin-RevId: a6ba4f82ef2bc1daec23c601b26ee814dc8afd3b