Remove gurchetansingh@chromium.org from third_party/minigbm/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ibb80d5cf920d4a6f05edbb118f788085c47c7673
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501955
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Michael Spang <spang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#979887}
NOKEYCHECK=True
GitOrigin-RevId: da70eb4ae0ab0445c6e6df4498d7a94a204a6fcc