[WebSQL] Add flag to disable WebSQL, and enterprise policy to override
WebSQL is a legacy standard, so there should add a way for users to
disable it and enterprises to re-enable if needed. Specifically:
(1) A new flag called web-sql-access, defaulting to enabled, that the
user can switch off.
(2) A new enterprise policy called WebSQLAccess, defaulting to disabled,
that can force enable WebSQL over the setting in the user's flag.

These are being added with an eye toward a future where WebSQL might be
disabled by default. When that's the case, the user would be able to
enable via the flag and/or the enterprise can enable via the policy.

TL;DR: if flag or policy is on, WebSQL is enabled. Enabled is still the
default state.

P.S. I'm not using the existing Database runtime feature because that's
hooked into WebPreferences. If Database is on, the flag and policy can
still disable WebSQL. If Database is off, the flag and policy have
no effect.

Bug: 695592
Change-Id: I7a48ef23179085dcbdc5fceec2bc3f8ca4e7b2c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3499154
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Ari Chivukula <arichiv@chromium.org>
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#979893}
NOKEYCHECK=True
GitOrigin-RevId: 8edcb697fe1619424c5780bfff1664ff6b641929