Create link highlight animation after verifying composited effect node.
Right now we create the link highlight animation whether or not the
node or effect node are painted. We should only create it after
verifying that the effect node has been composited and remove the
animation if the effect node is removed.

Bug: 900241
Change-Id: Ic1a92a28b3b64f15e6bf72f6f37a63aa373d64eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3517088
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#979945}
NOKEYCHECK=True
GitOrigin-RevId: a7c4dc01ef809543d75be89643e5b1b8ca507958