MF Frame Server Mode: Part 6: Media Engine Frame Server Enablement
This change introduces the switch for Media Engine to start in Frame
Server mode and then be promoted to Direct Composition mode except in
cases of protected content which will continue to start in Direct
Compositon mode.

A future change will enable the listening to the Promotion Hint Service
for promotion.

This also includes previous feedback that changes the callback for the
D3D11 video frame creation, removing the id token from the parameters
and passing it directly to the callback when created.

Bug: 1258887
Change-Id: I26040b034ed452ae4aadf7bd255a8159fdf2eb39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3433266
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: David Springgay <daspring@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#980241}
NOKEYCHECK=True
GitOrigin-RevId: ec10d2b7eca056741b9dfb6a57ac928f4aba04af