Fenced Frames: Create/destroy FencedFrameDelegate upon insertion/removal
This CL implements the changes outlined in
https://docs.google.com/document/d/1rbevNYwcX4jsr3a3L-1W96ffWNpGHSCJ0De9KzUQ4Eg/edit
which are required for introducing the `mode` IDL attribute.

Before this CL, HTMLFencedFrameElement had creation-time checks happen
in:
  HTMLFencedFrameElement::ctor()
    FencedFrameDelegate::Create()
      --> Creation-time checks

The creation-time checks determine whether or not we initialize the
`frame_delegate_` member. Making a one-time decision at creation-time as
to the fate of the `frame_delegate_` is too early, and leads to the
following problem:

 * Sandbox suitability was checked upon HTMLFencedFrameElement
   construction, meaning that if an HTMLFencedFrameElement was
   created in a hostile environment (and thus the FencedFrameDelegate
   was not created) and then adopted into a suitable environment, the
   frame would still be unusable. Before this CL, we fail a test for
   this [1]

Furthermore, it is confusing having all three of the following:
 1.) Creation-time checks: possibly leaving `frame_delegate_` null
 2.) Insertion-time checks: possible leaving `frame_delegate_` in a
     half-initialized state
 3.) Navigation-time checks

(1) gets us used to checking whether or not `frame_delegate_` is null
in various HTMLFencedFrameElement methods, but (2) is particularly
tricky to get wrong, as there is no good way to check if
`frame_delegate_` is partially-initialized. This leads to another test
that our implementation before this CL is failing [2].

After this CL, we remove (2) above, and move (1)'s creation-time checks
to be insertion-time checks. This means at insertion-time, we make the
decision as to whether or not we populate `frame_delegate_`.
Consequently, this CL removes FencedFrameDelegate::DidGetInserted(), and
its logic is moved into the FencedFrameDelegate constructor.

This CL also changes the lifetime of `frame_delegate_` to match the
underlying `content_frame_`. This has two effects:
 - FencedFrameShadowDOMDelegate: No longer lives as long as the
   ShadowRoot attached at the HTMLFencedFrameElement. Its ctor cannot
   assert that is creating the ShadowRoot for the very first time.
   Instead, it must just ensure [3] that it exists, since we never ever
   remove the ShadowRoot, as is customary in other HTML elements
 - FencedFrameMPArchDelegate: We move the DocumentFencedFrames
   registration (which increments the blink::Page::subframe_count_) to
   FFMPArchDelegate's ctor. Otherwise we'd have to conditionally call it
   in HTMLFencedFrameElement::DidNotifySubtreeInsertionsToDocument()
   depending on whether or not `frame_delegate_` was successfully
   populated. This is messier, and since the logic is MPArch-specific,
   we just move it to its delegate.

[1]:
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/wpt_internal/fenced_frame/create-in-sandbox-and-adopt-outside-sandbox.https.html
[2]:
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/wpt_internal/fenced_frame/maxframes.https.html;l=26-29;drc=a24831ddf0b711bf1d6d3a23e97aff2f8f2907a3
[3]:
https://chromium-review.googlesource.com/c/chromium/src/+/3501941/9/third_party/blink/renderer/core/html/fenced_frame/fenced_frame_shadow_dom_delegate.cc

R=dcheng@chromium.org

Bug: 1123606
Change-Id: I55e5d5e0be15ebcfe8f4abe462a63958a6b372cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501941
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Vladimir Nechaev <nechaev@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#980357}
NOKEYCHECK=True
GitOrigin-RevId: d3824397d79949344cc187b2c8b5894b1b348ba8