Set the time stamp of the synthetic mouse event and click
The other browsers dispatch click events before and after showing the
popup window, so we dispatch a mouse up and click events on the owner
element to match the legacy behavior. But we did not set the time stamp
of the synthetic events, which causes the event timing duration in
performance observer to be arbitrarily high. We should set the current
time stamp to these events.

Bug: 1230108
Change-Id: Ia8fe8a0ead908e24ee17bdd2b223e082fdcb91f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3518165
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#980787}
NOKEYCHECK=True
GitOrigin-RevId: 37b23a7aba94f9507c3613f9494acadc603615e9