[FlexNG] Multi-line column flex item expansion
As a result of CL:3290486, no special logic was needed to handle
expansion for items in a multi-line vs single-line column flex
container. Thus, add tests for the multi-line column case.

One bug with the iterator was uncovered while testing - in
NextLine() we weren't making use of next_item_idx_for_line_ and would
potentially re-traverse an item that had already been traversed with
its break token. This is now fixed.

Unrelatedly, multi-line-row-flex-fragmentation-016.html was updated
to adjust one of the OOF's offset and height to avoid overlap
with the previous row.

Bug: 660611
Change-Id: I92debb2d5c38b4d610219c387e74d476f7c10d7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3513328
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#980899}
NOKEYCHECK=True
GitOrigin-RevId: 987790716bd87d73e5d6343fc01bf6ba91e36743