Capability Delegation: fix payment token with local post-messaging
Moves the payment token from LocalFrame to LocalDOMWindow, and fixes
the token's activation call when the token-delegating post-message
call is sent to a same-origin Window.  Previously the activation call
was in a LocalFrame method that is used only by cross-origin
post-messages.

Fixed: 1302388
Change-Id: I01870785ba300514d984b0a6b5daefed635df9eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3503649
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Mustaq Ahmed <mustaq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981337}
NOKEYCHECK=True
GitOrigin-RevId: 2b308668693472425c208a7bbcb55d671ba0ff6f