[wasm] Transferring Memory.buffer should throw when not serialized
The ArrayBuffer that backs a WebAssembly.Memory object is
non-detachable, and as such it can't be transferred with the
structured clone algorithm. https://crrev.com/c/2653810 previously
fixed the case where such a buffer was transferred, but only in the
case where the buffer is reachable from the value to serialize.

Per the spec, transferring an object will detach the object and
perform any side effects of the transfer regardless of whether the
object has been serialized. And in the case of ArrayBuffer objects,
those side effects include throwing if the buffer is non-detachable.

Bug: 961059
Change-Id: Idaf5874ebf7ea42bce396166e7c075e74685ecd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3516873
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981455}
NOKEYCHECK=True
GitOrigin-RevId: 0b5e8598bf5df3e039b46da90f1c3618dd14fd1a